requirements.txt
pygame-ce>=2.5.0
pynetworktables>=2021.0.0
